Autotools for Qt part
