+++ .gitignore
@@ -27,11 +27,28 @@
 *.out
 *.app
 
-＃Interface
-*moc_*
-*.DS_Store
+# Project executable
+Casper
 
-# QtFile
+# Gedit family trash
+*~
+
+＃ Mac OSX trash
 .DS_Store
-Makefile
+
+# Qt trash
 moc_*
+
+# Autotools trash
+Makefile
+Makefile.in
+config.log
+config.status
+autom4te.cache/
+.deps/
+aclocal.m4
+config.guess
+config.sub
+depcomp
+install-sh
+missing
