+++ Casper.pro
@@ -1,12 +1,13 @@
 ######################################################################
-# Automatically generated by qmake (2.01a) ?? 11? 25 22:16:12 2014
+# Automatically generated by qmake (3.0) mer. nov. 26 15:06:16 2014
 ######################################################################
 
+QT += widgets
+
 TEMPLATE = app
-TARGET = 
-DEPENDPATH += . include/core src/core
-INCLUDEPATH += . include/core
+TARGET = Casper
+INCLUDEPATH += . include/gui
 
 # Input
-HEADERS += include/core/mainWindow.hpp
-SOURCES += src/core/main.cpp src/core/mainWindow.cpp
+HEADERS += include/gui/mainWindow.hpp
+SOURCES += src/gui/main.cpp src/gui/mainWindow.cpp
