+++ src/Makefile.am
@@ -1,0 +1,1 @@
+SUBDIRS = gui
